+++ Assets/Game/Scripts/MV/StatContext/Damage.cs
@@ -4,7 +4,7 @@ namespace Game.Scripts.MV.StatContext
 {
     public class Damage : Stat
     {
-        public Damage(float value, bool isPercentageValue) : base(value, isPercentageValue) { }
+        public Damage(int value, bool isPercentageValue) : base(value, isPercentageValue) { }
         
         protected override StatType GetStatType()
         {
