+++ Assets/Game/Scripts/MV/StatContext/Defence.cs
@@ -4,7 +4,7 @@ namespace Game.Scripts.MV.StatContext
 {
     public class Defence : Stat
     {
-        public Defence(float value, bool isPercentageValue) : base(value, isPercentageValue) { }
+        public Defence(int value, bool isPercentageValue) : base(value, isPercentageValue) { }
         
         protected override StatType GetStatType()
         {
